backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic[email]==2.10.5
pydantic-settings==2.7.1
langchain==0.3.27
langchain-core==0.3.81
langgraph==0.6.8
langgraph-checkpoint-postgres==2.0.25
psycopg[binary]==3.2.6
langchain-openai==0.3.7
langchain-google-genai==2.1.2
langchain-ollama==0.3.1
python-dotenv==1.0.1
httpx==0.28.1
PyPDF2==3.0.1
python-docx==1.1.2
qdrant-client==1.16.2
fastembed==0.6.1
beautifulsoup4==4.12.3
scikit-learn==1.5.2
